Improve default cherrypy json encoders
Use duck typing instead of inheritance for base Model and DisplayColor
Add comments
Consolidate into fewer files
