+++ pyproject.toml
@@ -16,6 +16,7 @@ ola = "^0.10.7"
 jinja2 = "^2.10"
 spectra = "^0.0.11"
 netifaces = "^0.10.9"
+orjson = "^3.5.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^3.0"
