+++ pyramidtriangles/shows/fuck_your_burn.py
@@ -4,10 +4,9 @@ from itertools import repeat, chain
 
 from ..color import HSV
 from ..grid import Position, Pyramid
-from .knobs import ValueKnob, HSVKnob
 from ..ponzicolor import Color, space
 from ..randomcolor import random_color
-from .show import Show
+from .show import Show, ValueKnob, HSVKnob
 
 
 class FuckYourBurn(Show):
